move demographics service layer to own file, move and update tests
